requirements.txt
openai~=0.25.0
pykanka~=0.1.0
PyYAML~=6.0